improving coord entering
improving the appearance and making it easier to enter the text
before there was "0" and "00.000" default in coord input. Now it's changed to HTML "placeholder" Attribute

Second thing is to automatically move to the next INPUT after reach maxlength.
Eg: when you enter "xy" to "lat_h" input you will be automaticly switched to "lat_min" input.